Move net.Conn logic to package main(conn)
